add tests for classes

+++ src/product.py
@@ -2,7 +2,7 @@ class Product:
     name: str
     description: str
     price: float
-    quantity: float
+    quantity: int
 
     def __init__(self, name, description, price, quantity):
         self.name = name
